backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
django-cors-headers==4.3.1
scikit-learn
pandas
numpy
joblib
firebase-admin
